Deduplicate `Principals` field
