+++ rs/nervous_system/proto/src/lib.rs
@@ -1,6 +1,6 @@
 use crate::pb::v1::Canister;
 use ic_base_types::PrincipalId;
-use pb::v1::{Decimal as DecimalPb, Duration, GlobalTimeOfDay, Percentage, Tokens};
+use pb::v1::{Decimal as DecimalPb, Duration, GlobalTimeOfDay, Percentage, Principals, Tokens};
 use rust_decimal::Decimal;
 use std::str::FromStr;
 
@@ -161,5 +161,16 @@ impl TryFrom<DecimalPb> for Decimal {
     }
 }
 
+impl From<Vec<PrincipalId>> for Principals {
+    fn from(principals: Vec<PrincipalId>) -> Self {
+        Self { principals }
+    }
+}
+
+impl From<Principals> for Vec<PrincipalId> {
+    fn from(principals: Principals) -> Self {
+        principals.principals
+    }
+}
 #[cfg(test)]
 mod tests;
