+++ rs/sns/swap/proto/ic_sns_swap/pb/v1/swap.proto
@@ -508,7 +508,7 @@ message CfNeuron {
   // with this neuron.
   uint64 amount_icp_e8s = 2;
   // The principals that can vote, propose, and follow on behalf of this neuron.
-  optional Principals hotkeys = 4;
+  optional ic_nervous_system.pb.v1.Principals hotkeys = 4;
 
   // Idempotency flag indicating whether the neuron recipes have been created for
   // the CfNeuron. When set to true, it signifies that the action of creating neuron
@@ -691,7 +691,7 @@ message CfInvestment {
   // The principal that can manage this neuron.
   optional ic_base_types.pb.v1.PrincipalId controller = 6;
   // The principals that can vote, propose, and follow on behalf of this neuron.
-  optional Principals hotkeys = 7;
+  optional ic_nervous_system.pb.v1.Principals hotkeys = 7;
 
   fixed64 nns_neuron_id = 2;
 
@@ -1171,12 +1171,6 @@ message SettleNeuronsFundParticipationRequest {
   message Aborted {}
 }
 
-// A list of principals.
-// Needed to allow prost to generate the equivalent of Optional<Vec<PrincipalId>>.
-message Principals {
-  repeated ic_base_types.pb.v1.PrincipalId principals = 1;
-}
-
 // Handling the Neurons' Fund and transferring some of its maturity to an SNS treasury is
 // thus the responsibility of the NNS Governance. When a swap succeeds, a Swap canister should send
 // a `settle_neurons_fund_participation` request to the NNS Governance, specifying its `result`
@@ -1199,7 +1193,7 @@ message SettleNeuronsFundParticipationResponse {
     // The principal that can manage this neuron.
     optional ic_base_types.pb.v1.PrincipalId controller = 6;
     // The principals that can vote, propose, and follow on behalf of this neuron.
-    optional Principals hotkeys = 7;
+    optional ic_nervous_system.pb.v1.Principals hotkeys = 7;
 
     // Whether the amount maturity amount of Neurons' Fund participation associated with this neuron
     // has been capped to reflect the maximum participation amount for this SNS swap.
